+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 I will do my best to guarantee that this project adheres to [Semantic Versioning](http://semver.org/) after 1.0.0, but please do read change log before updating.
 
+## Unreleased
+### Changed
+- `state` in `SocketProtocol` is changed to `status`.
+
 ## 0.8.1
 ### Changed
 - Fix versions of dependecies.